Use a single ReentrantMutex implementation on all platforms.
