+++ library/std/src/sys/windows/locks/mod.rs
@@ -2,5 +2,5 @@ mod condvar;
 mod mutex;
 mod rwlock;
 pub use condvar::{Condvar, MovableCondvar};
-pub use mutex::{MovableMutex, Mutex, ReentrantMutex};
+pub use mutex::{MovableMutex, Mutex};
 pub use rwlock::{MovableRwLock, RwLock};
